forging/forging/backend/requirements.txt
# ── FastAPI & Server ──
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9

# ── PyTorch & ML ──
torch>=2.2
torchvision>=0.17
segmentation-models-pytorch>=0.4.0
timm>=1.0.0

# ── Image Processing ──
numpy>=1.26
opencv-python-headless>=4.9
Pillow>=10.3
scipy>=1.13

# ── Document Handling ──
PyMuPDF>=1.24
imagehash>=4.3

# ── OCR ──
easyocr>=1.7

# ── Database ──
psycopg[binary]>=3.2

# ── Testing ──
pytest>=8.2
httpx>=0.27
